requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.7.2
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
crispy-bootstrap5==2024.2
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.2
distro==1.4.0
distro-info==0.23+ubuntu1.1
djando==0.2
Django==4.2.4
django-crispy-forms==2.1
django-import-export==3.3.1
django-seed==0.3.1
djangorestframework==3.15.1
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
Faker==19.3.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
GDAL==3.3.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupPy==1.14
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.3
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.2
pandas==2.0.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
platformdirs==2.0.2
protobuf==3.6.1
psycopg2-binary==2.9.7
pycairo==1.16.2
pycups==1.9.73
pygame==2.5.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
pytz==2023.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.4
suds-jurko==0.6
supervisor==4.1.0
systemd-python==234
tablib==3.5.0
toposort==1.10
typing-extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.6.0
virtualenv-clone==0.3.0
wadllib==1.3.3
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
